requirements.txt
langchain>=0.3.0
langchain-openai>=0.2.0
langchain-community>=0.3.0
langserve[all]>=0.3.0
faiss-cpu>=1.7.4
python-dotenv>=1.0.0
uvicorn[standard]>=0.30.0
fastapi>=0.115.0
beautifulsoup4>=4.12.0
